+++ src/app/delete.ts
@@ -3,13 +3,13 @@ import { exit } from 'node:process'
 import { intro, isCancel, log, outro, select } from '@clack/prompts'
 import { program } from 'commander'
 import { checkAppExistsAndHasPermissionOrgErr } from '../api/app'
-import { createSupabaseClient, findSavedKey, formatError, getConfig, getOrganizationId, OrganizationPerm, sendEvent, verifyUser } from '../utils'
+import { createSupabaseClient, findSavedKey, formatError, getAppId, getConfig, getOrganizationId, OrganizationPerm, sendEvent, verifyUser } from '../utils'
 
 export async function deleteApp(appId: string, options: OptionsBase) {
   intro(`Deleting`)
   options.apikey = options.apikey || findSavedKey()
   const extConfig = await getConfig()
-  appId = appId || extConfig?.config?.appId
+  appId = getAppId(appId, extConfig?.config)
 
   if (!options.apikey) {
     log.error('Missing API key, you need to provide a API key to upload your bundle')