+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [4.25.2](https://github.com/Cap-go/CLI/compare/v4.25.1...v4.25.2) (2024-11-12)
+
+
+### Bug Fixes
+
+* now use getAppId everywhere ([64acfa3](https://github.com/Cap-go/CLI/commit/64acfa381a3ba6d9e0e1e54c87d1a67bb649b79a))
+
+### [4.25.1](https://github.com/Cap-go/CLI/compare/v4.25.0...v4.25.1) (2024-11-12)
+
+
+### Bug Fixes
+
+* appId for broken cap config ([053fac5](https://github.com/Cap-go/CLI/commit/053fac513f1794e1ad6f3fa1539a78ebdc8b79cf))
+
+## [4.25.0](https://github.com/Cap-go/CLI/compare/v4.24.2...v4.25.0) (2024-11-12)
+
+
+### Features
+
+* node_modules flag ([6131cea](https://github.com/Cap-go/CLI/commit/6131cea2d9e2176643cfd86cb6ed34f07d7b1568))
+
 ### [4.24.2](https://github.com/Cap-go/CLI/compare/v4.24.1...v4.24.2) (2024-11-08)